tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/canonicalizeFilter/max-length-for-device-name-name.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';
test(function() {
    let name_too_long = 'a'.repeat(249);
    window.testRunner.setBluetoothMockDataSet('EmptyAdapter');
    assert_throws(new TypeError(), function () {
        window.navigator.bluetooth.requestDevice({filters: [{name: name_too_long}]});
    });
}, 'A device name longer than 248 must reject.');
</script>
